Move unit tests to appropriate module
